Implement ArrayNesting, AlienDictionary, FlattenMatrix, RangeAddition, TextJustification

+++ algorithms/java/src/Integer/BigInteger.java
@@ -1,4 +1,4 @@
-package BigInteger;
+package Integer;
 
 import java.util.Arrays;
 import java.util.LinkedList;
