+++ algorithms/java/test/Integer/BigIntegerTest.java
@@ -1,4 +1,4 @@
-package BigInteger;
+package Integer;
 
 import static junit.framework.TestCase.assertEquals;
 import static org.junit.Assert.assertArrayEquals;
